extract reusable files dealing with side effects to lib directory
