+++ app/lib/web3/BrowserWallet.ts
@@ -3,9 +3,9 @@ import hex2ascii = require("hex2ascii");
 import { injectable } from "inversify";
 import * as Web3 from "web3";
 
+import { WalletSubType, WalletType } from "../../modules/web3/types";
 import { EthereumAddress, EthereumNetworkId } from "../../types";
 import { IPersonalWallet, SignerType } from "./PersonalWeb3";
-import { WalletSubType, WalletType } from "./types";
 import { Web3Adapter } from "./Web3Adapter";
 
 export class BrowserWalletError extends Error {}
