+++ app/lib/web3/PersonalWeb3.ts
@@ -1,5 +1,5 @@
+import { WalletSubType, WalletType } from "../../modules/web3/types";
 import { EthereumAddress, EthereumNetworkId } from "../../types";
-import { WalletSubType, WalletType } from "./types";
 import { Web3Adapter } from "./Web3Adapter";
 
 // strings should match signer types on backend
