+++ app/modules/wallet-selector/ledger-wizard/flows.ts
@@ -3,15 +3,15 @@ import { toPairs, zip } from "lodash";
 import { pairZip } from "../../../../typings/modifications";
 import { GetState } from "../../../di/setupBindings";
 import { symbols } from "../../../di/symbols";
-import { injectableFn } from "../../../middlewares/redux-injectify";
-import { AppDispatch } from "../../../store";
-import { actions } from "../../actions";
 import {
   LedgerLockedError,
   LedgerNotAvailableError,
   LedgerWalletConnector,
-} from "../../web3/LedgerWallet";
-import { Web3Manager } from "../../web3/Web3Manager";
+} from "../../../lib/web3/LedgerWallet";
+import { Web3Manager } from "../../../lib/web3/Web3Manager";
+import { injectableFn } from "../../../middlewares/redux-injectify";
+import { AppDispatch } from "../../../store";
+import { actions } from "../../actions";
 import { walletFlows } from "../flows";
 
 export const LEDGER_WIZARD_SIMPLE_DERIVATION_PATHS = ["44'/60'/1'/0", "44'/60'/0'/0"]; // TODO this should be taken from config
