+++ test/components/walletSelector/WalletSelector.spec.tsx
@@ -9,16 +9,16 @@ import { BROWSER_WALLET_RECONNECT_INTERVAL } from "../../../app/components/walle
 import { LEDGER_RECONNECT_INTERVAL } from "../../../app/components/walletSelector/WalletLedgerInitComponent";
 import { WalletSelector } from "../../../app/components/walletSelector/WalletSelector";
 import { symbols } from "../../../app/di/symbols";
-import { obtainJwt } from "../../../app/modules/networking/jwt-actions";
 import {
   BrowserWallet,
   BrowserWalletConnector,
   BrowserWalletLockedError,
   BrowserWalletMissingError,
-} from "../../../app/modules/web3/BrowserWallet";
-import { LedgerWallet, LedgerWalletConnector } from "../../../app/modules/web3/LedgerWallet";
-import { Web3Adapter } from "../../../app/modules/web3/Web3Adapter";
-import { Web3Manager } from "../../../app/modules/web3/Web3Manager";
+} from "../../../app/lib/web3/BrowserWallet";
+import { LedgerWallet, LedgerWalletConnector } from "../../../app/lib/web3/LedgerWallet";
+import { Web3Adapter } from "../../../app/lib/web3/Web3Adapter";
+import { Web3Manager } from "../../../app/lib/web3/Web3Manager";
+import { obtainJwt } from "../../../app/modules/networking/jwt-actions";
 import { createMount } from "../../createMount";
 import { dummyNetworkId } from "../../fixtures";
 import {
